feat: make bbox.scale a tuple
Support x and y scaling, rotation, etc. by making BoundingBox.scale an
xy tuple. Now, any transformation matrix can be applied, including
matrices with xy scaling. There is no method to support this except
BoundingBox.transform with a matrix argument, but the support is now
there for flips and non-uniform scaling (where the bounding box will
still be a good fit) and even skew and rotation (for which the bounding
box will fit, but not tightly because it will still be locked into the x
and y axes by virtue of *being* a bounding box.

BREAKING CHANGE: Breaks any code that directly passes a value to
BoundingBox.scale