feat: use path i/o text obj in pad_text_ft
`pad_text_ft` returned a PaddedText instance where the `.elem` was a
"text" EtreeElement. `.elem` is not a "path" element. This produces a
smaller file than using Inkscape to convert text to path, because it
creates a path at the native (integer scale) size then transforms it
with a transformation matrix. The catch is "stroke-width", which scales
with the transformation matrix. `pad_text_ft` applies an inverse
transformation to "stroke-width" to anticipate this.