feat: use -y transformation for pad_text_ft PaddedText
`pad_text_ft` returns a PaddedText instance where the elem is a path.
This is inverted, because the splines in the font objects do not use
svg's "inverted" coordinate system. Init the bbox for these instances
with a -y transformation so the bbox transformation remains identical
(not just equivalent) to the element transform attrib.